requirements.txt
praw==7.7.1
python-dotenv==1.0.1
mistralai==0.1.6
pandas==2.3.1
httpx==0.25.2
orjson==3.11.0
pyarrow==15.0.2
